python/requirements.txt
azure-identity
azure-monitor-query
pandas